mcom03: Pass only low DDR regions to Linux
Both high and low DDR address ranges are currently available in Linux.
But GPU and EMACs do not have access to the high DDR address range
since GPU_BAR and HSPERIPH_BAR are set to 0x0 in U-Boot. Disable high
DDR address range in Linux as a temporary solution.

This commit limits the available memory in Linux to 2 GB.

Issue: #MCOM03BUP-234
Change-Id: I68c1f26a84278b57c4747470181c757e6709c3aa