phy: Add DQ mapping support
For LPDDR3/LPDDR4 CA training it is required to connect
PHY DQn lane to DRAM DQn for all n. On some MCom-03
boards this condition is not met. This commit allows to
map PHY DQn to DRAM DQm within each DBYTE, so the condition
for CA training is met.

DQ mapping is supported only for MCom-03 boards.

Issue: #MCOM03BUP-9
Change-Id: I76590efe0f2908123de726ed99e9c890a8f985d5