mcom03: defconfigs: Update impedance parameters
Previously, impedance parameters specified in defconfig were ignored
and the values hardcoded in the code were used. Now this is not the
case and impedance parameters in defconfig must be updated.

For MCom-03 BuB, ECAM03DM and Trustphone PM the same parameters are set
as they were previously hardcoded.

For ELV-MC03-SMARC new parameters are set suitable for ELV-MC03-SMARC
r1.0 and r1.1.

Issue: #MCOM03SW-1009
Issue: #MCOM03SW-1043
Change-Id: If2aa639a1b8d86bed0a7b73227d19a394ad95b4f