mcom03: Add DDR High region mapping
This commit adds mapping of 2 MB DDR High memory region to
32-bit address space accessible from RISC0. The mapping can
be used by bootloaders that run on RISC0 after ddrinit. For
example SBL uses it load TF-A image to DDR High.

It's convenient to do it in ddrinit, since ddrinit already
has VMMU driver. This eliminates the need to add VMMU driver
to the SBL and JTAG-script, which reduces code duplication.

Issue: #MCOM03SW-1410
Issue: #MCOM03SW-1623
Change-Id: Idfb468dea5d0e053f006fe4fec8c74567b427f6c