Add write of extended info of initialized DDR
We need to pass the data of initialized DDRMC's and dram size for
each individual DDRMC to CPU0 monitor. The existing solution
is passing memory regions data only and doesn't meet these
requirements.

This commit adds writing of extended info of initialized DDR
to be passed to U-Boot and CPU0 Monitor.

Issue: #ROB-270
Issue: #ROB-279
Depends-On: https://gerrit.elvees.com/c/bootloader/u-boot/+/33212
Change-Id: I1a288e9e8e2d4f862b5abc9bbb04fb59cefb05ca