mcom03: Move CONFIG_RAM_BASE addr to kseg0
Previously CONFIG_RAM_BASE pointed to kseg1 memory region, which is
uncacheable. Now we have pointed CONFIG_RAM_BASE to the cacheable kseg0
memory region that allows to reduce initialization time from ~900ms
to ~30ms.

The cache has to be enabled in an boot module which is responsible for
loading of ddrinit.

Issue: #MCOM03SW-1389
Change-Id: I1eceeb009902ee52e566af8f382100261f1d3883