ddrmc: lpddr4: Add constrains for row/column/banks address bits
According to DDRMC databook, MCom-03 does not support LPDDR4 16 Gb x8
(byte mode) SDRAMS. This means that maximum number of rows that DDRMC
can address is 131072 (17 address bits). The commit adds this limitation
to Kconfig.

The commit also adds column and bank constrains. According to LPDDR4
standard, number of columns is always 1024, number of banks is 8.

Change-Id: Ifc2bd4000a149dd41aaf529f6ba9309e8f9c5d08