lpddr4: Add support of C/A vref training
This is an optional training step that can be used to
optimize vref values for Command/Address lines.

This option is actually supported by both PHY firmware
versions (2023.09 and 2018.10), but it is undocumented
in 2018.10 version, so we have not used it.

The option is enabled by default for all LPDDR4 boards
except PM03CAM and PM03CAM_OSM based boards.

Issue: #MCOM03SW-3463
Change-Id: Ia1775565d0689937a4106f1eeb9012a5f0f5206e