Kconfig: Add CONFIG_DRAM_CAS_RDBI_ADDITIONAL_LAT option
When Read DBI is used DRAM CAS latency must be increased.
The commit adds new option that specifies how much it
must be increased depending on DRAM type and DRAM data rate.

Adding the option makes it easier to support Read DBI for DDR4,
that is added in next commit.

Issue: #MCOM03SW-3587
Change-Id: Ia852bd1fb0c8e3ad215ccd99eaef5edf162472e8