mcom03: Use LSP1 TIMERS0[7]
The LSP1 TIMERS0[7] is used and configured in SBL as system timer
starting from [1]. Thus the code responsible for the configuration
is removed.

Also the LSP1 subsystem initialization is done earlier in SBL (see [2]).
We want to have consistent settings of LSP1 through all boot stages
therefore we remove code responsible for the initialization of the
subsystem.

[1] https://gerrit.elvees.com/c/mcom03/sbl/+/55429
[2] https://gerrit.elvees.com/c/mcom03/sbl/+/51489

Issue: #MCOM03SW-3347
Depends-On: https://review.elvees.com/c/mcom03/linux/+/4688
Change-Id: Id3f7b872e57fcd7446556e5b73c166175898030e